+++ src/api/index.ts
@@ -1,5 +1,5 @@
 export * from "./constants.ts";
-export * from "./operations.ts";
+export * from "./DataStorageOperations.ts";
+export * from "./JSONMessageParts.ts";
 export * from "./packets";
-export * from "./parts.ts";
 export * from "./types.ts";
