+++ src/api/packets/client/LocationScouts.ts
