+++ src/api/packets/index.ts
@@ -1,52 +1,52 @@
-import { BouncePacket } from "./client/bounce.ts";
-import { ConnectPacket } from "./client/connect.ts";
-import { ConnectUpdatePacket } from "./client/connectUpdate.ts";
-import { GetPacket } from "./client/get.ts";
-import { GetDataPackagePacket } from "./client/getDataPackage.ts";
-import { LocationChecksPacket } from "./client/locationChecks.ts";
-import { LocationScoutsPacket } from "./client/locationScouts.ts";
-import { SayPacket } from "./client/say.ts";
-import { SetPacket } from "./client/set.ts";
-import { SetNotifyPacket } from "./client/setNotify.ts";
-import { StatusUpdatePacket } from "./client/statusUpdate.ts";
-import { SyncPacket } from "./client/sync.ts";
-import { BouncedPacket } from "./server/bounced.ts";
-import { ConnectedPacket } from "./server/connected.ts";
-import { ConnectionRefusedPacket } from "./server/connectionRefused.ts";
-import { DataPackagePacket } from "./server/dataPackage.ts";
-import { InvalidPacketPacket } from "./server/invalidPacket.ts";
-import { LocationInfoPacket } from "./server/locationInfo.ts";
-import { PrintJSONPacket } from "./server/printJSON.ts";
-import { ReceivedItemsPacket } from "./server/receivedItems.ts";
-import { RetrievedPacket } from "./server/retrieved.ts";
-import { RoomInfoPacket } from "./server/roomInfo.ts";
-import { RoomUpdatePacket } from "./server/roomUpdate.ts";
-import { SetReplyPacket } from "./server/setReply.ts";
+import { BouncePacket } from "./client/Bounce.ts";
+import { ConnectPacket } from "./client/Connect.ts";
+import { ConnectUpdatePacket } from "./client/ConnectUpdate.ts";
+import { GetPacket } from "./client/Get.ts";
+import { GetDataPackagePacket } from "./client/GetDataPackage.ts";
+import { LocationChecksPacket } from "./client/LocationChecks.ts";
+import { LocationScoutsPacket } from "./client/LocationScouts.ts";
+import { SayPacket } from "./client/SayPacket.ts";
+import { SetPacket } from "./client/Set.ts";
+import { SetNotifyPacket } from "./client/SetNotify.ts";
+import { StatusUpdatePacket } from "./client/StatusUpdate.ts";
+import { SyncPacket } from "./client/Sync.ts";
+import { BouncedPacket } from "./server/Bounced.ts";
+import { ConnectedPacket } from "./server/Connected.ts";
+import { ConnectionRefusedPacket } from "./server/ConnectionRefused.ts";
+import { DataPackagePacket } from "./server/DataPackage.ts";
+import { InvalidPacketPacket } from "./server/InvalidPacket.ts";
+import { LocationInfoPacket } from "./server/LocationInfo.ts";
+import { PrintJSONPacket } from "./server/PrintJSON.ts";
+import { ReceivedItemsPacket } from "./server/ReceivedItems.ts";
+import { RetrievedPacket } from "./server/Retrieved.ts";
+import { RoomInfoPacket } from "./server/RoomInfo.ts";
+import { RoomUpdatePacket } from "./server/RoomUpdate.ts";
+import { SetReplyPacket } from "./server/SetReply.ts";
 
-export * from "./client/bounce.ts";
-export * from "./client/connect.ts";
-export * from "./client/connectUpdate.ts";
-export * from "./client/get.ts";
-export * from "./client/getDataPackage.ts";
-export * from "./client/locationChecks.ts";
-export * from "./client/locationScouts.ts";
-export * from "./client/say.ts";
-export * from "./client/set.ts";
-export * from "./client/setNotify.ts";
-export * from "./client/statusUpdate.ts";
-export * from "./client/sync.ts";
-export * from "./server/bounced.ts";
-export * from "./server/connected.ts";
-export * from "./server/connectionRefused.ts";
-export * from "./server/dataPackage.ts";
-export * from "./server/invalidPacket.ts";
-export * from "./server/locationInfo.ts";
-export * from "./server/printJSON.ts";
-export * from "./server/receivedItems.ts";
-export * from "./server/retrieved.ts";
-export * from "./server/roomInfo.ts";
-export * from "./server/roomUpdate.ts";
-export * from "./server/setReply.ts";
+export * from "./client/Bounce.ts";
+export * from "./client/Connect.ts";
+export * from "./client/ConnectUpdate.ts";
+export * from "./client/Get.ts";
+export * from "./client/GetDataPackage.ts";
+export * from "./client/LocationChecks.ts";
+export * from "./client/LocationScouts.ts";
+export * from "./client/SayPacket.ts";
+export * from "./client/Set.ts";
+export * from "./client/SetNotify.ts";
+export * from "./client/StatusUpdate.ts";
+export * from "./client/Sync.ts";
+export * from "./server/Bounced.ts";
+export * from "./server/Connected.ts";
+export * from "./server/ConnectionRefused.ts";
+export * from "./server/DataPackage.ts";
+export * from "./server/InvalidPacket.ts";
+export * from "./server/LocationInfo.ts";
+export * from "./server/PrintJSON.ts";
+export * from "./server/ReceivedItems.ts";
+export * from "./server/Retrieved.ts";
+export * from "./server/RoomInfo.ts";
+export * from "./server/RoomUpdate.ts";
+export * from "./server/SetReply.ts";
 
 export type ClientPacket =
     | BouncePacket
