+++ src/api/packets/server/LocationInfo.ts
