+++ src/classes/managers/EventBasedManager.ts
@@ -1,3 +1,5 @@
+import { ArchipelagoEventEmitter } from "../ArchipelagoEventEmitter.ts";
+
 /**
  * An abstract class for managers that offer an event-based API.
  * @template Events An interface of events supported by this derived manager.
@@ -59,39 +61,5 @@ export abstract class EventBasedManager<Events extends { [p: string]: unknown[] 
      */
     protected emit<Event extends keyof Events & string>(event: Event, detail: Events[Event]): void {
         this.#events.dispatchEvent(event, detail);
-    }
-}
-
-/* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-argument */
-type EventCallback = (...args: any[]) => void;
-
-/**
- * A reimplementation of an EventEmitter that is runtime agnostic.
- * @internal
- */
-export class ArchipelagoEventEmitter {
-    #events: Record<string, [callback: EventCallback, once: boolean][]> = {};
-
-    public addEventListener(event: string, callback: EventCallback, once = false): void {
-        // Ensure list is not empty.
-        this.#events[event] ??= [];
-        this.#events[event].push([callback, once]);
-    }
-
-    public removeEventListener(event: string, callback: EventCallback): void {
-        const callbacks = this.#events[event];
-        if (callbacks && callbacks.length > 0) {
-            this.#events[event] = callbacks.filter(([cb]) => cb !== callback);
-        }
-    }
-
-    public dispatchEvent(event: string, data: any): void {
-        const callbacks = this.#events[event] ?? [];
-        for (const [callback, once] of callbacks) {
-            callback(...data);
-            if (once) {
-                this.removeEventListener(event, callback);
-            }
-        }
     }
 }
