+++ src/classes/managers/RoomStateManager.ts
@@ -1,13 +1,14 @@
-import { PermissionTable } from "../api";
-import { Client } from "../client.ts";
-import { EventBasedManager } from "./abstract.ts";
+import { PermissionTable } from "../../api";
+import { RoomStateEvents } from "../../events/RoomStateEvents.ts";
+import { ArchipelagoClient } from "../ArchipelagoClient.ts";
+import { EventBasedManager } from "./EventBasedManager.ts";
 
 /**
  * Managers room state information, notifies subscribers of changes, and exposes helper methods for interacting with the
  * room.
  */
 export class RoomStateManager extends EventBasedManager<RoomStateEvents> {
-    readonly #client: Client;
+    readonly #client: ArchipelagoClient;
     #serverVersion = { major: -1, minor: -1, build: -1 };
     #generatorVersion = { major: -1, minor: -1, build: -1 };
     #games: string[] = [];
@@ -121,11 +122,11 @@ export class RoomStateManager extends EventBasedManager<RoomStateEvents> {
     }
 
     /**
-     * Instantiates a new RoomStateManager. Should only be instantiated by creating a new {@link Client}.
+     * Instantiates a new RoomStateManager. Should only be instantiated by creating a new {@link ArchipelagoClient}.
      * @internal
      * @param client The client object this manager is associated with.
      */
-    public constructor(client: Client) {
+    public constructor(client: ArchipelagoClient) {
         super();
         this.#client = client;
 
@@ -196,51 +197,3 @@ export class RoomStateManager extends EventBasedManager<RoomStateEvents> {
             });
     }
 }
-
-/**
- * An interface with all supported room events and their respective callback arguments. To be called from
- * {@link RoomStateManager}.
- */
-export type RoomStateEvents = {
-    /**
-     * Fires when the room password has been toggled.
-     * @param password If the room now requires a password to join.
-     */
-    passwordUpdated: [password: boolean]
-
-    /**
-     * Fires when command permissions have been updated.
-     * @param oldValue The previous command permissions table.
-     * @param newValue The new command permissions table.
-     */
-    permissionsUpdated: [oldValue: PermissionTable, newValue: PermissionTable]
-
-    /**
-     * Fires when the location check points have been updated.
-     * @param oldValue The previous location check points value.
-     * @param newValue The new location check points value.
-     */
-    locationCheckPointsUpdated: [oldValue: number, newValue: number]
-
-    /**
-     * Fires when the hint cost has been updated.
-     * @param oldCost The previous amount of hint points required to request a hint.
-     * @param newCost The new amount of hint points required to request a hint.
-     * @param oldPercentage The old hint cost percentage.
-     * @param newPercentage The new hint cost percentage.
-     */
-    hintCostUpdated: [oldCost: number, newCost: number, oldPercentage: number, newPercentage: number]
-
-    /**
-     * Fires when the player's hint points value has updated.
-     * @param oldValue The old hint point value.
-     * @param newValue The new hint point value.
-     */
-    hintPointsUpdated: [oldValue: number, newValue: number]
-
-    /**
-     * Fires when new locations have been checked (or all locations on initial connection).
-     * @param locations All the newly checked locations.
-     */
-    locationsChecked: [locations: number[]]
-};
