+++ src/classes/managers/SocketManager.ts
@@ -1,28 +1,12 @@
-import {
-    BouncedPacket,
-    ClientPacket,
-    ConnectedPacket,
-    ConnectionRefusedPacket,
-    DataPackagePacket,
-    InvalidPacketPacket, JSONSerializableData,
-    LocationInfoPacket,
-    PrintJSONPacket,
-    ReceivedItemsPacket,
-    RetrievedPacket,
-    RoomInfoPacket,
-    RoomUpdatePacket,
-    ServerPacket,
-    SetReplyPacket,
-} from "../api";
-import { Client } from "../client.ts";
-import { EventBasedManager } from "./abstract.ts";
+import { ClientPacket, RoomInfoPacket, ServerPacket } from "../../api";
+import { SocketEvents } from "../../events/SocketEvents.ts";
+import { EventBasedManager } from "./EventBasedManager.ts";
 
 /**
  * Manages socket-level communication and exposes helper methods/events for interacting with the Archipelago API
  * directly.
  */
 export class SocketManager extends EventBasedManager<SocketEvents> {
-    #client: Client;
     #socket: WebSocket | null = null;
     #connected: boolean = false;
 
@@ -37,13 +21,11 @@ export class SocketManager extends EventBasedManager<SocketEvents> {
     }
 
     /**
-     * Instantiates a new SocketManager. Should only be instantiated by creating a new {@link Client}.
+     * Instantiates a new SocketManager. Should only be instantiated by creating a new {@link ArchipelagoClient}.
      * @internal
-     * @param client The client object this manager is associated with.
      */
-    public constructor(client: Client) {
+    public constructor() {
         super();
-        this.#client = client;
     }
 
     /**
@@ -102,7 +84,7 @@ export class SocketManager extends EventBasedManager<SocketEvents> {
 
         try {
             return new Promise((resolve, reject) => {
-                const IsomorphousWebSocket = findWebSocket();
+                const IsomorphousWebSocket = this.#findWebSocket();
                 if (IsomorphousWebSocket === null) {
                     throw new Error("Unable to find a suitable WebSocket API in the current runtime.");
                 }
@@ -212,144 +194,26 @@ export class SocketManager extends EventBasedManager<SocketEvents> {
             this.emit("receivedPacket", [packet]);
         }
     }
-}
 
-/**
- * An interface with all supported socket events and their respective callback arguments. To be called from
- * {@link SocketManager}.
- * @example
- * // Print all chat messages to the console when received.
- * client.socket.on("PrintJSON", (packet, message) => {
- *     console.log(message);
- * });
- *
- * // Warn when lost connection.
- * client.socket.on("Disconnect", () => {
- *     console.warn("Lost connection to the server!");
- * }
- */
-export type SocketEvents = {
-    /**
-     * Fires when the client receives a {@link BouncedPacket}.
-     * @param packet The raw {@link BouncedPacket}.
-     */
-    bounced: [packet: BouncedPacket, data: { [p: string]: JSONSerializableData }]
+    #findWebSocket(): typeof WebSocket | null {
+        let IsomorphousWebSocket: typeof WebSocket | null = null;
+        if (typeof window !== "undefined") {
+            // @ts-expect-error WebSocket may not exist in this context.
+            IsomorphousWebSocket = window.WebSocket || window.MozWebSocket;
+        } else if (typeof global !== "undefined") {
+            // @ts-expect-error WebSocket may not exist in this context.
+            IsomorphousWebSocket = global.WebSocket || global.MozWebSocket;
+        } else if (typeof self !== "undefined") {
+            // @ts-expect-error WebSocket may not exist in this context.
+            IsomorphousWebSocket = self.WebSocket || self.MozWebSocket;
+        } else if (typeof WebSocket !== "undefined") {
+            IsomorphousWebSocket = WebSocket;
+            // @ts-expect-error WebSocket may not exist in this context.
+        } else if (typeof MozWebSocket !== "undefined") {
+            // @ts-expect-error WebSocket may not exist in this context.
+            IsomorphousWebSocket = MozWebSocket as WebSocket;
+        }
 
-    /**
-     * Fires when the client receives a {@link ConnectedPacket}
-     * @param packet The raw {@link ConnectedPacket} packet.
-     * @remarks This also means the client has authenticated to an Archipelago server.
-     */
-    connected: [packet: ConnectedPacket]
-
-    /**
-     * Fires when the client receives a {@link ConnectionRefusedPacket}.
-     * @param packet The raw {@link ConnectionRefusedPacket}.
-     */
-    connectionRefused: [packet: ConnectionRefusedPacket]
-
-    /**
-     * Fires when the client receives a {@link DataPackagePacket}.
-     * @param packet The raw {@link DataPackagePacket}.
-     */
-    dataPackage: [packet: DataPackagePacket]
-
-    /**
-     * Fires when the client receives a {@link InvalidPacketPacket}.
-     * @param packet The raw {@link InvalidPacketPacket}.
-     */
-    invalidPacket: [packet: InvalidPacketPacket]
-
-    /**
-     * Fires when the client receives a {@link LocationInfoPacket}.
-     * @param packet The raw {@link LocationInfoPacket}.
-     */
-    locationInfo: [packet: LocationInfoPacket]
-
-    /**
-     * Fires when the client receives a {@link PrintJSONPacket}.
-     * @param packet The raw {@link PrintJSONPacket} packet.
-     * @param message The full plaintext message content.
-     */
-    printJSON: [packet: PrintJSONPacket, message: string]
-
-    /**
-     * Fires when the client receives a {@link ReceivedItemsPacket}.
-     * @param packet The raw {@link ReceivedItemsPacket}.
-     */
-    receivedItems: [packet: ReceivedItemsPacket]
-
-    /**
-     * Fires when the client receives a {@link RetrievedPacket}.
-     * @param packet The raw {@link RetrievedPacket}.
-     */
-    retrieved: [packet: RetrievedPacket]
-
-    /**
-     * Fires when the client receives a {@link RoomInfoPacket}.
-     * @param packet The raw {@link RoomInfoPacket}.
-     * @remarks This also means the client has established a websocket connection to an Archipelago server, but not yet
-     * authenticated.
-     */
-    roomInfo: [packet: RoomInfoPacket]
-
-    /**
-     * Fires when the client receives a {@link RoomUpdatePacket}.
-     * @param packet The raw {@link RoomUpdatePacket}.
-     */
-    roomUpdate: [packet: RoomUpdatePacket]
-
-    /**
-     * Fires when the client receives a {@link SetReplyPacket}.
-     * @param packet The raw {@link SetReplyPacket}.
-     */
-    setReply: [packet: SetReplyPacket]
-
-    /**
-     * Fires when the client receives any {@link ServerPacket}.
-     * @param packet Any received {@link ServerPacket}. Additional checks on the `cmd` property will be required to
-     * determine the type of packet received.
-     * @remarks All specific packet event listeners will fire before this event fires.
-     */
-    receivedPacket: [packet: ServerPacket]
-
-    /**
-     * Fires when the client sends an array of {@link ClientPacket}.
-     * @param packets An array of {@link ClientPacket} sent to the server.
-     */
-    sentPackets: [packets: ClientPacket[]]
-
-    /**
-     * Fires when the client has lost connection to the server, intentionally or not.
-     */
-    disconnected: []
-};
-
-/**
- * Finds the first appropriate WebSocket prototype available in the current scope.
- * @internal
- * @returns The first WebSocket class available or `null` if none available.
- * @remarks Not all features of the Web WebSocket API maybe available (due to runtime differences), so care should be
- * taken when interacting with the API directly.
- */
-function findWebSocket(): typeof WebSocket | null {
-    let IsomorphousWebSocket: typeof WebSocket | null = null;
-    if (typeof window !== "undefined") {
-        // @ts-expect-error WebSocket may not exist in this context.
-        IsomorphousWebSocket = window.WebSocket || window.MozWebSocket;
-    } else if (typeof global !== "undefined") {
-        // @ts-expect-error WebSocket may not exist in this context.
-        IsomorphousWebSocket = global.WebSocket || global.MozWebSocket;
-    } else if (typeof self !== "undefined") {
-        // @ts-expect-error WebSocket may not exist in this context.
-        IsomorphousWebSocket = self.WebSocket || self.MozWebSocket;
-    } else if (typeof WebSocket !== "undefined") {
-        IsomorphousWebSocket = WebSocket;
-        // @ts-expect-error WebSocket may not exist in this context.
-    } else if (typeof MozWebSocket !== "undefined") {
-        // @ts-expect-error WebSocket may not exist in this context.
-        IsomorphousWebSocket = MozWebSocket as WebSocket;
+        return IsomorphousWebSocket;
     }
-
-    return IsomorphousWebSocket;
 }
