+++ src/interfaces/ConnectionOptions.ts
@@ -1,50 +1,6 @@
-import { itemsHandlingFlags } from "./api";
-import { targetVersion } from "./constants.ts";
-import { generateUuid } from "./utils.ts";
-
-/**
- * An interface of client options that can be set on a {@link Client} object.
- */
-export interface ClientOptions {
-    /**
-     * The maximum number of milliseconds to wait for a response from the server when awaiting a response to a client
-     * packet.
-     * @default 10000
-     */
-    timeout?: number
-
-    /**
-     * Automatically requests the DataPackage from the server during {@link Client.login}, if the packages are missing
-     * in the {@link DataPackageManager}.
-     * @default true
-     */
-    autoFetchDataPackage?: boolean
-
-    /**
-     * Determines the maximum number of chat messages to log in {@link MessageManager}.
-     * @default 1000
-     * @remarks If `0` or fewer, message logging will effectively be disabled.
-     */
-    maximumMessages?: number
-
-    /**
-     * If enabled, logs the game, library version, and user agent to data storage, which can be used for debugging
-     * purposes.
-     * @default true
-     */
-    debugLogVersions: boolean
-}
-
-/**
- * Default {@link ClientOptions}.
- * @internal
- */
-export const defaultClientOptions: Required<ClientOptions> = {
-    timeout: 10000,
-    autoFetchDataPackage: true,
-    maximumMessages: 1000,
-    debugLogVersions: true,
-};
+import { itemsHandlingFlags } from "../api";
+import { targetVersion } from "../constants.ts";
+import { uuid } from "../utils.ts";
 
 /**
  * An interface of additional connection arguments when authenticating to an Archipelago server.
@@ -110,7 +66,7 @@ export interface ConnectionOptions {
  */
 export const defaultConnectionOptions: Required<ConnectionOptions> = {
     password: "",
-    uuid: generateUuid(),
+    uuid: uuid(),
     tags: [],
     version: targetVersion,
     items: itemsHandlingFlags.all,
