--- src/managers/index.ts
@@ -1,9 +1,0 @@
-export * from "./abstract.ts";
-export * from "./deathLink.ts";
-export * from "./items.ts";
-export * from "./message.ts";
-export * from "./package.ts";
-export * from "./players.ts";
-export * from "./room.ts";
-export * from "./socket.ts";
-export * from "./storage.ts";
