+++ src/utils.ts
@@ -2,7 +2,7 @@
  * Generate a random uuid version 4 hexadecimal string.
  * @internal
  */
-export function generateUuid(): string {
+export function uuid(): string {
     const uuid: (number | string)[] = [];
     for (let i = 0; i < 36; i++) {
         uuid.push(Math.floor(Math.random() * 16));
